backend-python/pyproject.toml
[project]
name = "scratch-backend"
version = "1.0.0"
description = "Scratch private deployment backend service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.3.0,<3.7.0",
    "beanie>=1.26.0,<2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "minio>=7.2.0",
    "redis>=5.2.0",
    "email-validator>=2.2.0",
    "litellm>=1.30.0",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
